Re-organize files in server directory. #4
