--- server/handle-error.js
@@ -1,5 +1,0 @@
-module.exports = function handleError(err, res, status, exit = 0) {
-  console.error(err)
-  res.sendStatus(status)
-  if (!exit) process.exit(exit)
-}
